requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
alembic==1.13.2
asgiref==3.8.1
attrs==24.2.0
blinker==1.8.2
bootstrap4==0.1.0
click==8.1.7
Django==5.1
django-crispy-forms==2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Flask==3.0.3
Flask-Admin==1.6.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
greenlet==3.0.3
home==0.3.4
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.0.5
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.9.0
pyserial==3.5
python-dateutil==2.9.0.post0
redis==5.0.8
rfxcom==0.5.0
simplejson==3.19.3
six==1.16.0
SQLAlchemy==2.0.32
sqlparse==0.5.1
typing_extensions==4.12.2
uWSGI==2.0.26
Werkzeug==3.0.3
WTForms==3.1.2
yarl==1.9.4
